Fix tests and implmentation
